requirements.txt
firebase-admin==6.5.0
pandas==2.2.2
requests==2.32.3
streamlit==1.36.0
Flask==3.0.3
Pyrebase4==4.8.0
streamlit-autorefresh==1.0.0
